fix(apply issue): markdown loading
